requirements.txt
Django==1.7.3
django-appconf==0.6
django-compressor==1.3
django-model-utils==2.0
django-realtime==1.1
django-user-sessions==1.0.0
gunicorn==18.0
mgrs==1.1.0
networkx==1.8.1
psycopg2==2.5.2
python-dateutil==2.2
redis==2.10.3
six==1.5.2
South==0.8.4
xlrd==0.9.2
